create utils folder and refactor functions
