+++ src/utils/generateIndex.js
@@ -1,0 +1,4 @@
+export default function generateIndex() {
+	let index = (Math.random().toPrecision(5) * 100000).toString();
+	return index;
+}
