+++ src/utils/index.js
@@ -1,0 +1,13 @@
+import generateIndex from './generateIndex'
+import setImageOnCanvas from './setImageOnCanvas'
+import putFileInStorage from './putFileInStorage'
+import loadingProgress from './loadingProgress'
+import getCanvasAndContext from './getCanvasAndContext'
+
+export {
+    generateIndex,
+    setImageOnCanvas,
+    putFileInStorage,
+    loadingProgress,
+    getCanvasAndContext
+}
